Create an remove book event delegation on the container
The event delegation is placed on the container that
holds all of the book cards (divs) and accesses
the remove buttons when one of the buttons is clicked.
When a remove button is clicked, the event delegation
targets the button that was clicked and looks at the
parent element. It searches for the data attribute of
the parent element, in this case bookId. We can then
get an index of bookId by searching for a matching
value. Once the value is found we can remove that element
based on the index from the array and the DOM.