Changed the logic of the read status functionality
Created a protoype on the Book object called toggleReadStatus.
This function simply changes the read status from false to true and
vice versa. This caused me to change the event delegation that was assigned
to the container element for when the readStatusButton is clicked.
It is much simpler now though as the event delegation simply calls the
toggleReadStatus prototype function.